// sdk/node/src/generated/grpc/tradeapi/v1/auth/auth_service.ts
export interface AuthRequest {
  /** API token (secret key) */
  secret: string;
  /** Request source application ID */
  sourceAppId: string;
}

/** Authorization information */
export interface AuthResponse {
  /** Received JWT token */
  token: string;
}

/** Requesting information about a token */
export interface TokenDetailsRequest {
  /** JWT token */
  token: string;
}

/** Token Information */
export interface TokenDetailsResponse {
  /** Creation date and time */
  createdAt:
    | Date
    | undefined;
  /** Expiration date and time */
  expiresAt:
    | Date
    | undefined;
  /** Information about accessing market data */
  mdPermissions: MDPermission[];
  /** Account IDs */
  accountIds: string[];
  /** Session and trading accounts in the token will be marked readonly */
  readonly: boolean;
}
// ...
export interface SubscribeJwtRenewalRequest {
  /** API token (secret key) */
  secret: string;
  /** Request source application ID */
  sourceAppId: string;
}

/** Updated token. Stream */
export interface SubscribeJwtRenewalResponse {
  /** Received JWT token */
  token: string;
}

function createBaseAuthRequest(): AuthRequest {
  return { secret: "", sourceAppId: "" };
}

export const AuthRequest: MessageFns<AuthRequest> = {
  encode(message: AuthRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.secret !== "") {
      writer.uint32(10).string(message.secret);
    }
    if (message.sourceAppId !== "") {
      writer.uint32(18).string(message.sourceAppId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AuthRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAuthRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.secret = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.sourceAppId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AuthRequest {
    return {
      secret: isSet(object.secret) ? globalThis.String(object.secret) : "",
      sourceAppId: isSet(object.sourceAppId)
        ? globalThis.String(object.sourceAppId)
        : isSet(object.source_app_id)
        ? globalThis.String(object.source_app_id)
        : "",
    };
  },

  toJSON(message: AuthRequest): unknown {
    const obj: any = {};
    if (message.secret !== "") {
      obj.secret = message.secret;
    }
    if (message.sourceAppId !== "") {
      obj.sourceAppId = message.sourceAppId;
    }
    return obj;
  },

  create(base?: DeepPartial<AuthRequest>): AuthRequest {
    return AuthRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<AuthRequest>): AuthRequest {
    const message = createBaseAuthRequest();
    message.secret = object.secret ?? "";
    message.sourceAppId = object.sourceAppId ?? "";
    return message;
  },
};

function createBaseAuthResponse(): AuthResponse {
  return { token: "" };
}

export const AuthResponse: MessageFns<AuthResponse> = {
  encode(message: AuthResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.token !== "") {
      writer.uint32(10).string(message.token);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AuthResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAuthResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.token = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AuthResponse {
    return { token: isSet(object.token) ? globalThis.String(object.token) : "" };
  },

  toJSON(message: AuthResponse): unknown {
    const obj: any = {};
    if (message.token !== "") {
      obj.token = message.token;
    }
    return obj;
  },

  create(base?: DeepPartial<AuthResponse>): AuthResponse {
    return AuthResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<AuthResponse>): AuthResponse {
    const message = createBaseAuthResponse();
    message.token = object.token ?? "";
    return message;
  },
};

function createBaseTokenDetailsRequest(): TokenDetailsRequest {
  return { token: "" };
}

export const TokenDetailsRequest: MessageFns<TokenDetailsRequest> = {
  encode(message: TokenDetailsRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.token !== "") {
      writer.uint32(10).string(message.token);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TokenDetailsRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTokenDetailsRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.token = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TokenDetailsRequest {
    return { token: isSet(object.token) ? globalThis.String(object.token) : "" };
  },

  toJSON(message: TokenDetailsRequest): unknown {
    const obj: any = {};
    if (message.token !== "") {
      obj.token = message.token;
    }
    return obj;
  },

  create(base?: DeepPartial<TokenDetailsRequest>): TokenDetailsRequest {
    return TokenDetailsRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TokenDetailsRequest>): TokenDetailsRequest {
    const message = createBaseTokenDetailsRequest();
    message.token = object.token ?? "";
    return message;
  },
};

function createBaseTokenDetailsResponse(): TokenDetailsResponse {
  return { createdAt: undefined, expiresAt: undefined, mdPermissions: [], accountIds: [], readonly: false };
}

export const TokenDetailsResponse: MessageFns<TokenDetailsResponse> = {
  encode(message: TokenDetailsResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.createdAt !== undefined) {
      Timestamp.encode(toTimestamp(message.createdAt), writer.uint32(10).fork()).join();
    }
    if (message.expiresAt !== undefined) {
      Timestamp.encode(toTimestamp(message.expiresAt), writer.uint32(18).fork()).join();
    }
    for (const v of message.mdPermissions) {
      MDPermission.encode(v!, writer.uint32(26).fork()).join();
    }
    for (const v of message.accountIds) {
      writer.uint32(34).string(v!);
    }
    if (message.readonly !== false) {
      writer.uint32(40).bool(message.readonly);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TokenDetailsResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTokenDetailsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.createdAt = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.expiresAt = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.mdPermissions.push(MDPermission.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.accountIds.push(reader.string());
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.readonly = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TokenDetailsResponse {
    return {
      createdAt: isSet(object.createdAt)
        ? fromJsonTimestamp(object.createdAt)
        : isSet(object.created_at)
        ? fromJsonTimestamp(object.created_at)
        : undefined,
      expiresAt: isSet(object.expiresAt)
        ? fromJsonTimestamp(object.expiresAt)
        : isSet(object.expires_at)
        ? fromJsonTimestamp(object.expires_at)
        : undefined,
      mdPermissions: globalThis.Array.isArray(object?.mdPermissions)
        ? object.mdPermissions.map((e: any) => MDPermission.fromJSON(e))
        : globalThis.Array.isArray(object?.md_permissions)
        ? object.md_permissions.map((e: any) => MDPermission.fromJSON(e))
        : [],
      accountIds: globalThis.Array.isArray(object?.accountIds)
        ? object.accountIds.map((e: any) => globalThis.String(e))
        : globalThis.Array.isArray(object?.account_ids)
        ? object.account_ids.map((e: any) => globalThis.String(e))
        : [],
      readonly: isSet(object.readonly) ? globalThis.Boolean(object.readonly) : false,
    };
  },

  toJSON(message: TokenDetailsResponse): unknown {
    const obj: any = {};
    if (message.createdAt !== undefined) {
      obj.createdAt = message.createdAt.toISOString();
    }
    if (message.expiresAt !== undefined) {
      obj.expiresAt = message.expiresAt.toISOString();
    }
    if (message.mdPermissions?.length) {
      obj.mdPermissions = message.mdPermissions.map((e) => MDPermission.toJSON(e));
    }
    if (message.accountIds?.length) {
      obj.accountIds = message.accountIds;
    }
    if (message.readonly !== false) {
      obj.readonly = message.readonly;
    }
    return obj;
  },

  create(base?: DeepPartial<TokenDetailsResponse>): TokenDetailsResponse {
    return TokenDetailsResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TokenDetailsResponse>): TokenDetailsResponse {
    const message = createBaseTokenDetailsResponse();
    message.createdAt = object.createdAt ?? undefined;
    message.expiresAt = object.expiresAt ?? undefined;
    message.mdPermissions = object.mdPermissions?.map((e) => MDPermission.fromPartial(e)) || [];
    message.accountIds = object.accountIds?.map((e) => e) || [];
    message.readonly = object.readonly ?? false;
    return message;
  },
};
// ...
export interface AuthServiceClient<CallOptionsExt = {}> {
  /**
   * Exchange an API secret for a JWT
   * Example HTTP request:
   * POST /v1/sessions
   * Content-Type: application/json
   *
   * {
   *   "secret": "your-api-secret-key"
   * }
   *
   * All fields are sent in the request body
   */
  auth(request: DeepPartial<AuthRequest>, options?: CallOptions & CallOptionsExt): Promise<AuthResponse>;
  /**
   * Get session token information
   * Example HTTP request:
   * POST /v1/sessions/details
   * Content-Type: application/json
   *
   * {
   *   "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
   * }
   *
   * The token is sent in the request body for security
   * The response also includes the accounts available to the token.
   */
  tokenDetails(
    request: DeepPartial<TokenDetailsRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<TokenDetailsResponse>;
  /** Subscription to JWT token update. Stream method */
  subscribeJwtRenewal(
    request: DeepPartial<SubscribeJwtRenewalRequest>,
    options?: CallOptions & CallOptionsExt,
  ): AsyncIterable<SubscribeJwtRenewalResponse>;
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;